added forecast icon
