+++ Geo-Weather/Geo-Weather/Resources/Assets.xcassets/Icons/sunny.imageset/clear.png
